Fri, Jul 10, 2026, 1:27 PM +05:30

+++ Week 1/calculator.c
@@ -3,7 +3,7 @@
 
 int main(void)
 {
-   int dollars = 1;
+   long dollars = 1;
    while (true)
    {
     char c = get_char("Here's $%li. Double it and give to the next person ? ",dollars);
